Merge in SV/sv from SVN-2103-Update_Gitian_Build to develop

* commit 'ca1038f715e3c09818cb7800fce6b07dd7f4f632': (23 commits)
  GCC 12 has changed uninitialised warning catagories
  Fix deprecated std::iterator warning
  Fix array subscript wwarning in CBloomFilter constructor
  Fx clang warning for implicit long to double conversion
  Fix clang warning for ambiguous operator==
  Fix clang build error for incomplete type CCancellationSource
  Update CMake build
  Fix signed comparison warnings with latest boost.
  Fix deprecated boost::filesystem warnings
  Deprecated use of volatile
  std::accumulate now moves its lhs argument
  Fix signal/slots and bind errors
  Fix C++20 compilation errors
  Update build instructions for supported platforms.
  LibTBB checking and linking
  Configure for c++20
  Update autoconf CXX version detection script.
  Fix symbol versioning
  Bump g++ to version 10
  Bump version to focal
  ...

+++ src/rpc/net.cpp
@@ -225,7 +225,7 @@ static UniValue getpeerinfo(const Config &config,
         if (stats.dPingTime > 0.0) {
             obj.push_back(Pair("pingtime", stats.dPingTime));
         }
-        if (stats.dMinPing < std::numeric_limits<int64_t>::max() / 1e6) {
+        if (stats.dMinPing < static_cast<double>(std::numeric_limits<int64_t>::max()) / 1e6) {
             obj.push_back(Pair("minping", stats.dMinPing));
         }
         if (stats.dPingWait > 0.0) {
